% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_fn}
\alias{app_fn}
\title{Create a \code{\link[shiny:shinyApp]{shinyApp}} object}
\usage{
app_fn(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link[shiny:shinyApp]{shinyApp()}}}
}
\value{
A \code{\link[shiny:shinyApp]{shinyApp}} object.
}
\description{
Create a \code{shinyApp} object for \code{analyzeBehaviorspace}
}
\examples{
# ADD_EXAMPLES_HERE
}
