% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{classify_vars}
\alias{classify_vars}
\title{Classify BehaviorSpace experiment variables into independent and dependent.}
\usage{
classify_vars(df)
}
\arguments{
\item{df}{The data frame from a BehaviorSpace experiment.}
}
\value{
A named list containing character vectors of independent and
dependent variables.
}
\description{
Classify variables in a BehaviorSpace experiment data frame into
dependent and independent. This function uses the fact that BehaviorSpace
puts the independent variable columns before the \verb{[steps]} column, and the
dependent variables after.
}
