% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{extract_last_tick}
\alias{extract_last_tick}
\title{Extract the rows of a data frame with the last tick from each run}
\usage{
extract_last_tick(df)
}
\arguments{
\item{df}{The data frame.}
}
\value{
A data frame with one row for each run, corresponding to the last
tick for that run.
}
\description{
Extract the rows of a data frame with the last tick from each run.
}
