% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{get_group_vars}
\alias{get_group_vars}
\title{Get the possible grouping variables from an experiment}
\usage{
get_group_vars(experiment, x_var, y_var)
}
\arguments{
\item{experiment}{An experiment object, returned by \code{\link[=load_bs_file]{load_bs_file()}},
\code{\link[=load_bs_table]{load_bs_table()}}, or \code{\link[=load_bs_spreadsheet]{load_bs_spreadsheet()}}.}

\item{x_var}{The name of the x-variable to use in the plot.}

\item{y_var}{The name of the y-variable to use in the plot.}
}
\value{
A data frame with the names of the possible grouping variables
and the corresponding columns in the experiment data frame.
}
\description{
Given x- and y-variables to plot, get a list of the possible variables to
use in grouping the data for plotting. These variables are the independent
variables other than the y-variable.
}
