% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{get_plot_vars}
\alias{get_plot_vars}
\title{Get the variables from an experiment that will not be used in a plot.}
\usage{
get_plot_vars(experiment, x_var, y_var)
}
\arguments{
\item{experiment}{An experiment object, returned by \code{\link[=load_bs_file]{load_bs_file()}},
\code{\link[=load_bs_table]{load_bs_table()}}, or \code{\link[=load_bs_spreadsheet]{load_bs_spreadsheet()}}.}

\item{x_var}{The name of the x-variable to use in the plot.}

\item{y_var}{The name of the x-variable to use in the plot.}
}
\value{
A data frame with the names of the independent variables that will
not be plotted.
}
\description{
Given x- and y-variables to plot, get a list of the other independent
variables that will not be used in the plot.
}
