% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{get_yvars}
\alias{get_yvars}
\title{Get the possible y-variables from an experiment}
\usage{
get_yvars(experiment, x_var)
}
\arguments{
\item{experiment}{An experiment object, returned by \code{\link[=load_bs_file]{load_bs_file()}},
\code{\link[=load_bs_table]{load_bs_table()}}, or \code{\link[=load_bs_spreadsheet]{load_bs_spreadsheet()}}.}

\item{x_var}{The name of the x-variable to use in the plot.}
}
\value{
A data frame with the names of the possible y-variables
and the corresponding columns in the experiment data frame.
}
\description{
Given an x-variable to plot, get a list of the possible y-variables.
}
