% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{is_bs_table}
\alias{is_bs_table}
\title{Determine whether a csv file is a BehaviorSpace experiment table}
\usage{
is_bs_table(text, skip_lines)
}
\arguments{
\item{text}{The contents of the .csv file.}

\item{skip_lines}{The number of lines to skip before starting to read the
data.}
}
\value{
Logical: \code{TRUE} means it is in the table format.
}
\description{
Given the text from a .csv file, determine whether it corresponds to a
the table format for a BehaviorSpace experiment
}
