% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_analyze_behaviorspace.R
\name{launch_abs}
\alias{launch_abs}
\title{Launch the AnalyzeBehaviorspace app}
\usage{
launch_abs(rstudio = getOption("analyze_behaviorspace.rstudio"), ...)
}
\arguments{
\item{rstudio}{Only relevant for RStudio users. The default (\code{FALSE}) is
to launch the app in the user's default web browser rather than RStudio's
pop-up Viewer. Users can change the default to \code{TRUE} by setting the
global option \code{options(shinystan.rstudio = TRUE)}.}

\item{...}{Optional arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
Nothing is returned
}
\description{
Launch the AnalyzeBehaviorspace app in the default web browser. RStudio
users also have the option of launching the app in RStudio's pop-up Viewer.
}
\examples{
\dontrun{
launch_abs()
}

}

