% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{process_bs_data}
\alias{process_bs_data}
\alias{load_bs_spreadsheet}
\alias{load_bs_file}
\title{Process data from a BehaviorSpace experiment}
\usage{
process_bs_data(d, quiet = TRUE)

load_bs_spreadsheet(filename = NULL, text = NULL, quiet = TRUE)

load_bs_file(filename = NULL, text = NULL, quiet = TRUE)
}
\arguments{
\item{d}{A data frame containing experiment data read from a .csv file.}

\item{quiet}{Logical value indicating whether to run quietly or report
progress messages.}
}
\value{
A named list with elements:
* \code{data}: a data frame containing the experiment data.
* \code{ind_vars}: a character vector with the names of the independent
  variables.
* \code{dep_vars}: a character vector with the names of the dependent
  variables.
* \code{mapping}: a data frame mapping columns in \code{data} to variable
  names. By default, this just maps column names to themselves.
* \code{success}: A logical variable indicating success or failure.
* \code{cause}: A character variable indicating the cause of failure.
}
\description{
Take the data read from a spreadsheet or tabular BehaviorSpace experiment
and process it
}
\details{
This function loads and decodes output from a NetLogo BehaviorSpace
experiment. It can take either a filename or the text contents of such a
file.
}
\section{Functions}{
\itemize{
\item \code{process_bs_data}: Load a BehaviorSpace experiment in table format.

\item \code{load_bs_spreadsheet}: Load a BehaviorSpace experiment in spreadsheet
format.
}}

