% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plot.R
\name{make_plot}
\alias{make_plot}
\title{Make a ggplot from a BehaviorSpace experiment}
\usage{
make_plot(
  experiment,
  points,
  lines,
  x_var,
  y_var,
  group_var,
  error_bars = "none",
  last_tick = FALSE
)
}
\arguments{
\item{experiment}{A BehaviorSpace experiment object, as returned from
\code{\link[=load_bs_file]{load_bs_file()}}, \code{\link[=load_bs_table]{load_bs_table()}}, or \code{\link[=load_bs_experiment]{load_bs_experiment()}}.}

\item{points}{Logical: plot points.}

\item{lines}{Logical: plot lines.}

\item{x_var}{The name of the x-variable.}

\item{y_var}{The name of the y-variable.}

\item{group_var}{Optionally, the name of a grouping variable.}

\item{error_bars}{The kind of error bars to use. Legal values are
"none", "error bars", or "bands"}

\item{last_tick}{Logical: Plot only the last tick from each run.}
}
\value{
A ggplot object.
}
\description{
Make a ggplot plot from a BehaviorSpace experiment, using a chosen x-variable,
y-variable, and optional grouping variable. In addition, the user may
specify error bars for the plot, and may choose to plot all ticks, or only
the last tick of each run.
}
