% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{process_bs_data}
\alias{process_bs_data}
\title{Process data from a BehaviorSpace \code{.csv} file.}
\usage{
process_bs_data(d, quiet = TRUE)
}
\arguments{
\item{d}{A data frame read from a BehaviorSpace .csv file.}

\item{quiet}{Logical value indicating whether to run quietly or report
progress messages.}
}
\value{
A named list with elements:
\itemize{
\item \code{data}: a data frame containing the experiment data.
\item \code{ind_vars}: a character vector with the names of the independent
variables.
\item \code{dep_vars}: a character vector with the names of the dependent
variables.
\item \code{mapping}: a data frame mapping columns in \code{data} to variable
names. By default, this just maps column names to themselves.
\item \code{success}: A logical variable indicating success or failure.
\item \code{cause}: A character variable indicating the cause of failure.
}
}
\description{
\code{process_bs_data()} takes a data frame read from a \code{.csv} file with
BehaviorSpace output and organizes it into an R representation of
a BehaviorSpace experiment.
}
\details{
This function is generally not run by the user, but is called internally
by \code{\link[=load_bs_file]{load_bs_file()}}, \code{\link[=load_bs_table]{load_bs_table()}}, or \code{\link[=load_bs_spreadsheet]{load_bs_spreadsheet()}}.
}
\seealso{
\code{\link[=load_bs_file]{load_bs_file()}}, \code{\link[=load_bs_table]{load_bs_table()}}, and \code{\link[=load_bs_spreadsheet]{load_bs_spreadsheet()}}.
}
