% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_fn}
\alias{server_fn}
\title{Create a \code{\link[shiny:shinyServer]{shinyServer}} object.}
\usage{
server_fn()
}
\value{
A \code{\link[shiny:shinyServer]{shinyServer}} object.
}
\description{
Create a \code{\link[shiny:shinyServer]{shinyServer}} object to use in the
\code{analyzeBeahviorspace} \code{\link[shiny:shinyApp]{shinyApp}}.
}
\examples{
\dontrun{
ui <- ui_fn()
server <- server_fn()
shinyApp(ui, server)
}

}
