% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{tx_name}
\alias{tx_name}
\title{Translate a text name to a column name in the experiment data frame}
\usage{
tx_name(var_name, mapping)
}
\arguments{
\item{var_name}{The name of a variable.}

\item{mapping}{A mapping data frame from an experiment.}
}
\value{
The name of the data frame column corresponding to the variable name.
}
\description{
Take a text name (e.g., something a variable has been renamed to in
AnalyzeBehaviorSpace) and translate it to a column name in the experiment
data frame.
}
