% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_fn}
\alias{ui_fn}
\title{Create a \code{\link[shiny:shinyUI]{shinyUI}} object}
\usage{
ui_fn()
}
\value{
A \code{\link[shiny:shinyUI]{shinyUI}} object.
}
\description{
Create a \code{\link[shiny:shinyUI]{shinyUI}} object to use in a
\code{analyzeBehaviorspace} \code{\link[shiny:shinyApp]{shinyApp}}.
}
\examples{
\dontrun{
ui <- ui_fn()
server <- server_fn()
shinyApp(ui, server)
}

}
