% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_behaviorspace_packages.R
\docType{package}
\name{analyzeBehaviorspace}
\alias{analyzeBehaviorspace}
\alias{analyzeBehaviorspace-package}
\title{analyzeBehaviorspace: Processes NetLogo BehaviorSpace Output}
\description{
Reads in table-style output from NetLogo BehaviorSpace experiments, allows the user to do some basic organizing and interactive plotting, and to export charts and summary tables.
}
\section{License}{


The \pkg{analyzeBehaviorspace} package is open source licensed under the
MIT License.
}

\section{Bug reports}{

\itemize{
\item analyzeBehaviorspace issue tracker (\url{https://github.com/jonathan-g/analyzeBehaviorspace/issues})
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathan-g/analyzeBehaviorspace}
  \item \url{https://jonathan-g.github.io/analyzeBehaviorspace}
  \item Report bugs at \url{https://github.com/jonathan-g/analyzeBehaviorspace/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Gilligan \email{jonathan.gilligan@vanderbilt.edu} (\href{https://orcid.org/0000-0003-1375-6686}{ORCID})

}
\keyword{internal}
