% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_util.R
\name{get_expt_vars}
\alias{get_expt_vars}
\title{Get the variables from an experiment}
\usage{
get_expt_vars(experiment)
}
\arguments{
\item{experiment}{An experiment object, returned by [load_bs_file()],
[load_bs_table()], or [load_bs_spreadsheet()].}
}
\value{
A data frame with the names of the variables and the corresponding
  columns in the experiment data frame.
}
\description{
Get the variables from a BehaviorSpace experiment.
}
