% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_plot.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Get data frame for plotting from an experiment}
\usage{
get_plot_data(experiment, x_var, y_var, group_var, last_tick)
}
\arguments{
\item{experiment}{An experiment object, returned by [load_bs_file()],
[load_bs_table()], or [load_bs_spreadsheet()].}

\item{x_var}{The name of the independent variable for the x-axis}

\item{y_var}{The name of the dependent variable for the y-axis.}

\item{group_var}{The name of the variable to use for grouping.}

\item{last_tick}{Logical: plot only the last tick (as opposed to plotting all
ticks).}
}
\value{
A data frame suitable for plotting with `ggplot()`.
}
\description{
From an experiment, extract a data frame suitable for plotting with
`ggplot()`. Select the relevant columns for plotting, using an x-variable,
a y-variable, and an optional grouping variable. If `last_tick` is `TRUE`,
then only include data for the last tick of each run. Otherwise, include all
ticks.
}
