% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{load_bs_file}
\alias{load_bs_file}
\alias{load_bs_table}
\alias{load_bs_spreadsheet}
\title{Load output from a NetLogo BehaviorSpace experiment}
\usage{
load_bs_file(filename = NULL, text = NULL, quiet = TRUE)

load_bs_table(filename = NULL, text = NULL, quiet = TRUE)

load_bs_spreadsheet(filename = NULL, text = NULL, quiet = TRUE)
}
\arguments{
\item{filename}{The name of a BehaviorSpace table output file (\code{.csv}
format).}

\item{text}{Text contents of a BehaviorSpace table output file (\code{.csv}
format).}

\item{quiet}{Logical value indicating whether to run quietly or report
progress messages.}
}
\value{
A named list with elements:
\itemize{
\item \code{data}: a data frame containing the experiment data.
\item \code{ind_vars}: a character vector with the names of the independent
variables.
\item \code{dep_vars}: a character vector with the names of the dependent
variables.
\item \code{mapping}: a data frame mapping columns in \code{data} to variable
names. By default, this just maps column names to themselves.
\item \code{success}: A logical variable indicating success or failure.
\item \code{cause}: A character variable indicating the cause of failure.
}
}
\description{
Loads the output from a NetLogo BehaviorSpace zexperiment.
}
\details{
This function loads and decodes table output from a NetLogo BehaviorSpace
experiment. It can take either a filename or the text contents of such a
file.
}
\section{Functions}{
\itemize{
\item \code{load_bs_table}: Load a BehaviorSpace experiment in table
format.

\item \code{load_bs_spreadsheet}: Load a BehaviorSpace experiment in spreadsheet
format.
}}

\examples{
\dontrun{
  load_bs_file(
    system.file("test_data/butterfly_small-experiment-table.csv",
      "analyzeBehaviorspace", mustWork = TRUE)
    )
}

}
