% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abs_load_data.R
\name{load_bs_table}
\alias{load_bs_table}
\alias{process_bs_data}
\alias{load_bs_spreadsheet}
\alias{load_bs_file}
\title{Load output from a NetLogo BehaviorSpace experiment}
\usage{
load_bs_table(filename = NULL, text = NULL, quiet = TRUE)

process_bs_data(d, quiet = TRUE)

load_bs_spreadsheet(filename = NULL, text = NULL, quiet = TRUE)

load_bs_file(filename = NULL, text = NULL, quiet = TRUE)
}
\arguments{
\item{filename}{The name of a BehaviorSpace table output file (\code{.csv}
format).}

\item{text}{Text contents of a BehaviorSpace table output file (\code{.csv}
format).}

\item{quiet}{Logical value indicating whether to run quietly or report
progress messages.}

\item{d}{A data frame read from a BehaviorSpace .csv file.}
}
\value{
A named list with elements:
* \code{data}: a data frame containing the experiment data.
* \code{ind_vars}: a character vector with the names of the independent
  variables.
* \code{dep_vars}: a character vector with the names of the dependent
  variables.
* \code{mapping}: a data frame mapping columns in \code{data} to variable
  names. By default, this just maps column names to themselves.
* \code{success}: A logical variable indicating success or failure.
* \code{cause}: A character variable indicating the cause of failure.
}
\description{
\code{load_bs_table} loads the table output from a NetLogo BehaviorSpace
experiment.
}
\details{
This function loads and decodes table output from a NetLogo BehaviorSpace
experiment. It can take either a filename or the text contents of such a
file.
}
\section{Functions}{
\itemize{
\item \code{load_bs_table}: Load a BehaviorSpace experiment in spreadsheet
format.

\item \code{process_bs_data}: Load a BehaviorSpace experiment in table format.

\item \code{load_bs_spreadsheet}: Load a BehaviorSpace experiment in spreadsheet
format.
}}

\examples{
\dontrun{
  load_bs_table(
    system.file("test_data/butterfly_small-experiment-table.csv",
      "analyzeBehaviorspace", mustWork = TRUE)
    )
}

}
